third_party/blink/web_tests/fast/dom/Window/window-postmessage-clone-frames-expected.txt

Tests that we clone object hierarchies
PASS: (frames[0].tests[0]).eventData instanceof (frames[0]).Object is true of type boolean
PASS: (frames[0].tests[0]).event.data instanceof (frames[0]).Object is true of type boolean
PASS: (frames[0].tests[1]).eventData instanceof (frames[0]).Array is true of type boolean
PASS: (frames[0].tests[1]).event.data instanceof (frames[0]).Array is true of type boolean
PASS: (frames[0].tests[2]).eventData instanceof (frames[0]).Date is true of type boolean
PASS: (frames[0].tests[2]).event.data instanceof (frames[0]).Date is true of type boolean
PASS: (frames[0].tests[3]).eventData instanceof (frames[0]).File is true of type boolean
PASS: (frames[0].tests[3]).event.data instanceof (frames[0]).File is true of type boolean
PASS: (frames[0].tests[4]).eventData instanceof (frames[0]).FileList is true of type boolean
PASS: (frames[0].tests[4]).event.data instanceof (frames[0]).FileList is true of type boolean
